+++ CHANGELOG.md
@@ -380,5 +380,8 @@ Unreleased
 * Custom ini keys that are added to the [Basic] and [Map] sections, and to any of the House sections, are now preserved, and editable in the "INI Rules & Tweaks" section. This will allow passive support for modded features.
 * Previews are now shown for Celltrigger placement.
 * Fixed errors in the detection of Aftermath expansion units on maps.
-* Fixed collision detection of multi-cell objects to detect crossing outside the map.
+* Fixed map load collision detection of multi-cell objects to detect crossing outside the map.
 * Fixed the NoOwnedObjectsInSole option to also correctly toggle loading and saving of these objects.
+* Added option in image export to only export the map bounds. This is enabled by default for multiplayer maps.
+* The Map Objects Statistics tool now shows the classic game limits for TD maps. (RA has no changes in these.)
+* Disabled the ability to export to .meg type via the menu shortcut when the editor is not compiled in developer mode.